+++ lib/resource/toolsFastor/network/NetworkTypeDio.dart
@@ -3,6 +3,7 @@ enum NetworkTypeDio{
   get,
   post,
   put,
+  patch,
   delete,
   file
 }