+++ lib/resource/toolsFastor/network/dio/DioService.dart
@@ -177,6 +177,61 @@ extension DioService on NetworkManagerDio {
 
   }
 
+  //--------------------------------------------------------------------------- patch
+
+  Future<Response>  patch_dio() async {
+    Response? response ;
+    try {
+      // FormData form = FormData.fromMap(body);
+
+      Dio _dio = Dio();
+
+      //show request and response in beatful log
+      if( isEnableLogDioPretty! ) {
+        _dio.interceptors.add(PrettyDioLogger(
+          requestHeader: isEnableLogDioPretty!,
+          requestBody: isEnableLogDioPretty!,
+          responseBody: isEnableLogDioPretty!,
+        ));
+      }
+
+
+      response =   await _dio.patch(url, options: Options(headers: headers), data: body);
+      // Log.k(tag, "_put() - success: " + response.toString());
+
+      //call back
+      if (callback_dio != null) callback_dio!(true, "success", response.data);
+
+      //return
+      return response;
+    } on DioError catch (dioError) {
+      Log.k(tag, "DioError - e: " + dioError.toString());
+      if( dioError.response != null && dioError!.response!.data != null ) {
+        Log.k(tag, "DioError - statusCode: " + dioError!.response!.statusCode.toString());
+        Map<String,dynamic> data = Map();
+        if( dioError!.response!.data != null ) {
+          if( dioError!.response!.data is Map ) {
+            data = dioError!.response!.data;
+          } else {
+            data["data"] = dioError!.response!.data.toString();
+          }
+        }
+        if (callback_dio != null) callback_dio!(false, "failed", data);
+        return dioError!.response!;
+      } else {
+        if (callback_dio != null) callback_dio!(false, "failed",  Map() );
+        return getFailedResponse(msg: dioError.toString() );
+      }
+
+    } catch (e) {
+      String msg = e.toString();
+      Log.k(tag, "_put() - e: " + msg);
+      if (callback_dio != null) callback_dio!(false, msg, Map());
+      return getFailedResponse(msg: e.toString() );
+    }
+
+  }
+
   //---------------------------------------------------------------------------- file
 
   Future<Response>  file_dio() async {